all_practices_for_learning/homino-project/requirements.txt
pip install django

# بررسی نصب
django-admin --version